Logger, refactor & bugfixes - [alert-worker] added logger - [alert-worker] refactoring. - [alert-worker] fix on empty cases (no face_log) - [alert-worker] reworked alert-generator and location-update to run on its own
