+++ .gitignore
@@ -11,6 +11,7 @@ node_modules
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
+**.log
 
 # Editor directories and files
 .idea
@@ -25,7 +26,6 @@ yarn-error.log*
 !.gitkeep
 
 # server stuff
-server/logs/**.log
 server/public/images/uploads/*
 
 !server/public/images/uploads/.gitkeep
