+++ alert-worker/package.json
@@ -5,7 +5,7 @@
   "main": "src/main.js",
   "scripts": {
     "start": "npm run dev",
-    "dev": "node src/main.js",
+    "dev": "cross-env NODE_ENV=development node src/main.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "",
@@ -15,9 +15,11 @@
     "dotenv": "^8.2.0",
     "lodash": "^4.17.15",
     "luxon": "^1.22.0",
-    "mysql2": "^2.1.0"
+    "mysql2": "^2.1.0",
+    "winston": "^3.2.1"
   },
   "devDependencies": {
+    "cross-env": "^7.0.2",
     "faker": "^4.1.0",
     "why-is-node-running": "^2.1.2"
   }
