+++ alert-worker/src/alert-types.js
@@ -1,0 +1,5 @@
+module.exports = {
+  UNKNOWN: 'U',
+  UNAUTHORIZED: 'A',
+  OVERSTAY: 'O'
+}
