--- alert-worker/src/test.js
@@ -1,16 +1,0 @@
-const mysql = require('mysql2/promise')
-const _ = require('lodash')
-require('dotenv').config()
-
-const pool = mysql.createPool({
-  host: process.env.DB_HOST,
-  port: process.env.DB_PORT || 3306,
-  database: process.env.DB_NAME,
-  user: process.env.DB_USER,
-  password: process.env.DB_PASS,
-  multipleStatements: true,
-  waitForConnections: true,
-  connectionLimit: 10,
-  queueLimit: 2
-})
-
